requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
idna==3.3
jsonschema==4.17.3
Pillow==9.3.0
psycopg2-binary==2.9.5
pyproject-toml==0.0.10
pyrsistent==0.19.2
python-decouple==3.6
pytz==2022.6
requests==2.28.1
simplejson==3.18.0
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.7
urllib3==1.26.13
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.1
Django==4.2.16
django-phonenumber-field==5.2.0
django-phonenumbers==1.0.1
phonenumbers==8.13.44
pycparser==2.22
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.1
typing_extensions==4.12.2
